improve 'started' message with proper cause
Before, the started message would say the build had been 'started
by changes ...' if a user started a build when there were changes
to be pulled in.

Now if a user starts the build, it will always say 'started by user',
but if a user didn't start the build, it will still show the changes.