search upstream causes when gathering commit list
The "r.hasChangeSetComputed()" check is superfluous and looks wrong, so remove it:
http://javadoc.jenkins-ci.org/hudson/model/AbstractBuild.html#hasChangeSetComputed()

Before, if a downstream change was triggered by a commit and the
"Show Commit List with Titles and Authors" option was enabled, the downstream
change would notify "No Changes.".
Now, the downstream change will notify with the list of changes from an
upstream cause that has changes. If there are no upstream changes, it
still notifies "No Changes.".