Refs #129 Duration in back to normal message is now the duration between successful builds
Added constants for build status messages.
Back to normal status is now given if the previous build was unstable or a failure.
Back to normal status is now only given if any previous builds for the project have been success, if this is the firt time a build for the project has passed then just show success status.
Append duration method now calls another method to calculate the duration between successful builds if the status is back to normal.