Run our migration process before any other
In order to fix the compatibility issue with EnvInject plugin,
we need to ensure that our migration process runs before the one
in the EnvInject plugin. Extensions are sorted in the descending order
of the ordinal, so setting ordinal = 100 will make our extension first
in the list returned by ItemListener.all()

This ordering approach is fragile and usually discouraged, but this
should work as the ordinal of the EnvInject migration extension is 0
(which is the default value).