samrocketman no longer maintainer
I've tried to effectively find someone who is willing to go through this plugin and lead on its design.  As of now, I'm no longer maintaining it and am officially dropping my maintenance of it.  If you wish, step up and actively maintain this plugin to drive its direction and development.  Maintaining this has been an interesting experience for me considering I don't use the plugin in practice.  I feel it would be in much better hands for someone with drive.

I apologize if this inconveniences any of you but I currently wish to focus my time elsewhere.

Thanks,
SAM