Add multiple channel comment to slackSend docs (#399)
The global plugin configuration where the comment about accepting a csv
string for channels is only rendered in the plugin documentation. This
changes adds that useful piece of knowledge to the slackSend step
documentation like in https://jenkins.io/doc/pipeline/steps/slack/ where
the global config is not easily available.