Add reply_broadcast option to StandardSlackService
This allows posting messages into threads while still broadcasting them to the
channel as well. This allows using threads but still notifying the channel in
case of serious errors or other situations requiring attention.

[Slack documentation][1] says:

> Provide a `thread_ts` value for the posted message to act as a reply to
a parent message. Sparingly set `reply_broadcast` to `true` if your reply is
important enough for everyone in the channel to receive.

[1]: https://api.slack.com/methods/chat.postMessage#threads_and_replies