Make bot messages respect the message parameter (#552)
When calling StandardSlackService.publish(String message, JSONArray attachments, String color) as a bot-user the message parameter is silently dropped. I'm not sure if this is a bug or intentional.

I was trying to send plain text messages (without the attachments colour) when I spotted this by using slackSend with an empty attachments array.