Follow the established exception handling convention.
The general established behaviours in BuildStep is to simply let the
caller handle the exception, possibly with exception chaining to add
more context.

What the code does right now is to simply swallow the stack trace part
of the exception, which is often the most crucial, especially for
unanticipated exceptions (in this context, there's no checked exceptions
thrown.)

See issues like https://github.com/jenkinsci/slack-plugin/issues/595
where this behaviour makes it painful for users (and maintainers) to
troubleshoot what's going on.